+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.9.6
 - Fixed Portable Grid losing enchantments when placing and breaking (raoulvdberge)
+- Added client config option to remember the Grid search query (raoulvdberge)
 
 ### 1.9.5
 - Re-added the `/refinedstorage disk create <player> <id>` command (raoulvdberge)